food-delivery foods app microservice boldi

+++ back-end/foods/views/favourite_views.py
@@ -8,7 +8,7 @@ from authen.renderers import UserRenderers
 from foods.models import Foods, Favorite
 from utils.pagination import StandardResultsSetPagination
 from utils.pagination import Pagination
-from utils.user_permission import check_user_permission
+from utils.permission_auth import check_user_permission
 from foods.serializers.favourite_serializers import (
     FoodsSerializer,
     FavoritesSerializer,
@@ -22,11 +22,14 @@ class FoodCategories(APIView, Pagination):
     pagination_class = StandardResultsSetPagination
     serializer_class = FoodsSerializer
 
-    def get(self, request, pk, format=None, *args, **kwargs):
+    @check_user_permission
+    def get(self, request, pk, user_id=None, format=None, *args, **kwargs):
+        if user_id is None:
+            return Response({"error": "Invalid user data"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
         instance = Foods.objects.filter(categories=pk)
         page = super().paginate_queryset(instance)
         if page is not None:
-            serializer = super().get_paginated_response(self.serializer_class(page, many=True, context={"user": request.user, "request": request}).data)
+            serializer = super().get_paginated_response(self.serializer_class(page, many=True, context={"user": user_id, "request": request}).data)
         else:
             serializer = self.serializer_class(instance, many=True)
         return Response({"data": serializer.data}, status=status.HTTP_200_OK)
@@ -39,8 +42,10 @@ class FavouritesView(APIView, Pagination):
     serializer_class = FavoritesSerializer
 
     @check_user_permission
-    def get(self, request):
-        queryset = Favorite.objects.filter(user=request.user.id, is_favorite=True)
+    def get(self, request, user_id=None):
+        if user_id is None:
+            return Response({"error": "Invalid user data"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        queryset = Favorite.objects.filter(user=user_id, is_favorite=True)
         page = super().paginate_queryset(queryset)
         if page is not None:
             serializer = super().get_paginated_response(self.serializer_class(page, many=True, context={"request": request}).data)
@@ -50,8 +55,10 @@ class FavouritesView(APIView, Pagination):
 
     @check_user_permission
     @swagger_auto_schema(request_body=FavoriteSerializer)
-    def post(self, request):
-        serializers = FavoriteSerializer(data=request.data, context={"user": request.user})
+    def post(self, request, user_id=None):
+        if user_id is None:
+            return Response({"error": "Invalid user data"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        serializers = FavoriteSerializer(data=request.data, context={"user": user_id})
         if serializers.is_valid(raise_exception=True):
             serializers.save()
             return Response(serializers.data, status=status.HTTP_201_CREATED)
@@ -63,7 +70,9 @@ class FavouriteViews(APIView):
     perrmisson_class = [IsAuthenticated]
 
     @check_user_permission
-    def delete(self, request, pk):
+    def delete(self, request, pk, user_id=None):
+        if user_id is None:
+            return Response({"error": "Invalid user data"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
         objects_get = Favorite.objects.get(food=pk)
         objects_get.delete()
         return Response({"message": "Delete success"}, status=status.HTTP_200_OK)
